requirements.txt
earthengine-api
geemap
sentence-transformers
faiss-cpu
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-groq
python-dotenv
streamlit
streamlit-chat